+++ packages/client/src/components/fileManager/index.tsx
@@ -1,4 +1,5 @@
 import { ImageFile } from '@metastable/types';
+import clsx from 'clsx';
 import React, { useState } from 'react';
 
 import { FileList } from './FileList';
@@ -9,6 +10,7 @@ interface Props {
   onOpen?: (itemIds: string[]) => void;
   actions?: JSX.Element;
   selectionActions?: (selection: ImageFile[]) => JSX.Element;
+  className?: string;
 }
 
 export const FileManager: React.FC<Props> = ({
@@ -16,11 +18,12 @@ export const FileManager: React.FC<Props> = ({
   onOpen,
   actions,
   selectionActions,
+  className,
 }) => {
   const [selection, setSelection] = useState<string[]>([]);
 
   return (
-    <div className={styles.manager}>
+    <div className={clsx(styles.manager, className)}>
       <div className={styles.actions}>
         {actions}
         {selection.length > 0 && (
@@ -34,12 +37,16 @@ export const FileManager: React.FC<Props> = ({
           </>
         )}
       </div>
-      <FileList
-        items={items}
-        onOpen={onOpen}
-        selection={selection}
-        onSelect={setSelection}
-      />
+      {items.length ? (
+        <FileList
+          items={items}
+          onOpen={onOpen}
+          selection={selection}
+          onSelect={setSelection}
+        />
+      ) : (
+        <div className={styles.info}>No files of this type are available.</div>
+      )}
     </div>
   );
 };